pyproject.toml
[project]
name = "rageval-nba"
version = "0.1.0"
description = "Evaluation harness for hybrid RAG systems, with an NBA analytics reference implementation."
authors = [{name = "Daniel Keinan", email = "dk3425@columbia.edu"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "rich>=13.9.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "sqlite-vec>=0.1.0",
    "nba-api>=1.7.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "types-pyyaml",
]

[project.scripts]
rageval = "rageval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rageval"]
artifacts = ["src/rageval/templates/*.j2", "src/rageval/examples/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RET"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = ["ignore::pytest.PytestCollectionWarning"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
